organization and global parameter linking
