+++ qtgui/__init__.py
@@ -2,9 +2,7 @@ __author__ = "Hiranmayi Duvvuri"
 __date__ = "2017-01-08"
 __all__ = []
 
-from . import exp_widgets
-
 from .exp_setup import *
 from .main_window import *
 from .exp_frames import *
-from .exp_widgets import *
+from .exp_sliders import *
