+++ qtgui/exp_sliders.py
@@ -1,0 +1,316 @@
+from qtpy.QtGui import *
+from qtpy.QtCore import *
+from qtpy.QtWidgets import *
+
+#from .indiv_exp import *
+
+class SlidersExpanded(QWidget):
+	"""
+	extra slider options pop-up
+	"""
+
+	def __init__(self, exp, fitter):
+		super().__init__()
+
+		self._exp = exp
+		self._fitter = fitter
+
+		self.layout()
+
+	def layout(self):
+
+		main_layout = QGridLayout(self)
+
+		# update bounds, link???
+
+
+class Sliders(QWidget):
+	"""
+	create sliders for an experiment
+	"""
+
+	def __init__(self, exp, param_name, fitter, global_exp):
+		super().__init__()
+
+		self._exp = exp
+		self._param_name = param_name
+		self._fitter = fitter
+		self._global_exp = global_exp
+
+		self.layout()
+
+	@property
+	def param_name(self):
+
+		return self._param_name
+
+	def bounds(self):
+		"""
+		"""
+		pass
+
+
+	def layout(self):
+		"""
+		"""
+		self._layout = QGridLayout(self)
+		self._layout.setVerticalSpacing(40)
+
+		self._name_label = QLabel(self._param_name, self)
+		self._layout.addWidget(self._name_label, 0, 0, 0, 2)
+
+		self._fix = QCheckBox("Fix?", self)
+		self._fix.toggle()
+		self._fix.setChecked(False)
+		self._fix.stateChanged.connect(self.fix)
+		self._layout.addWidget(self._fix, 1, 0)
+
+		self._slider = QSlider(Qt.Horizontal)
+		self._slider.valueChanged[int].connect(self.update_val)
+		self._layout.addWidget(self._slider, 1, 1)
+
+		self.bounds()
+
+		self._fix_int = QLineEdit(self)
+		self._layout.addWidget(self._fix_int, 1, 2)
+		self._fix_int.setText(str(1))
+		self._fix_int.hide()
+
+		self._expand = QPushButton("...", self)
+		self._expand.clicked.connect(self.expanded)
+		self._layout.addWidget(self._expand, 1, 4)
+
+	def expanded(self):
+
+		self._more = SlidersExpanded(self._exp, self._fitter)
+		self._more.setGeometry(500, 400, 400, 100)
+		self._more.show()
+
+	def fix(self, state):
+
+		if state == Qt.Checked:
+			# change widget views
+			self._fix_int.show()
+			self._slider.hide()
+
+			self._fitter.fix(self._exp, **{self._param_name: int(self._fix_int.text())})
+			print(self._fix_int.text())
+		else:
+			#change widget views
+			self._fix_int.hide()
+			self._slider.show()
+
+			self._fitter.unfix(*[self._param_name], expt = self._exp)
+			print('unfixed')
+
+	def update_val(self, value):
+
+		self._fitter.update_guess(self._param_name, value, self._exp)
+		print(value)
+
+
+class LocalSliders(Sliders):
+	"""
+	create sliders for an experiment
+	"""
+
+	def __init__(self, exp, param_name, value, fitter, global_list, slider_list, global_exp):
+
+		self._global_list = global_list
+		self._slider_list = slider_list
+		self._value = value
+
+		super().__init__(exp, param_name, fitter, global_exp)
+
+	def bounds(self):
+
+		exp_range = self._exp.model.param_guess_ranges[self._param_name]
+
+		self._slider.setMinimum(exp_range[0])
+		self._slider.setMaximum(exp_range[1])
+
+		self._link = QComboBox(self)
+		self._link.addItem("Unlink")
+		self._link.addItem("Add Global Var")
+
+		for i in self._global_list:
+			self._link.addItem(i)
+
+		self._link.activated[str].connect(self.link_unlink)
+		self._layout.addWidget(self._link, 1, 3)
+
+	def link_unlink(self, status):
+
+		if status == "Unlink":
+			self._fitter.unlink_from_global(self._exp, self._param_name)
+			self._global_exp[status].unlinked(self)
+			print("unlinked")
+		elif status == "Add Global Var":
+			text, ok = QInputDialog.getText(self, "Add Global Variable", "Var Name: ")
+			if ok: 
+				self._global_list.append(text)
+				for e in self._slider_list["Local"].values():
+					for i in e:
+						i.update_global(text)
+		else:
+			self._fitter.link_to_global(self._exp, self._param_name, status)
+			self._slider.hide()
+			self._fix.hide()
+
+			if status not in self._global_exp:
+				self._global_exp[status] = GlobalExp(self._fitter, status, self._slider_list, self._global_list, self._global_exp)
+
+			self._global_exp[status].linked(self)
+			print("linked to " + status)
+
+	def update_global(self, value):
+
+		self._link.addItem(value)
+
+	def reset(self):
+
+		self._slider.show()
+		self._fix.show()
+
+class GlobalSliders(Sliders):
+	"""
+	create sliders for an experiment
+	"""
+
+	def __init__(self, exp, param_name, fitter, global_exp):
+		super().__init__(exp, param_name, fitter, global_exp)
+
+	def bounds(self):
+
+		exp_range = self._fitter.param_ranges[0][self._param_name]
+
+		self._slider.setMinimum(exp_range[0])
+		self._slider.setMaximum(exp_range[1])
+		
+
+class Experiments(QWidget):
+	"""
+	experiment box widget
+	"""
+
+	def __init__(self, fitter, name, slider_list, global_var, global_exp):
+		super().__init__()
+
+		self._fitter = fitter
+		self._name = name
+		self._slider_list = slider_list
+		self._global_var = global_var
+		self._global_exp = global_exp
+
+		self.layout()
+
+	def layout(self):
+		main_layout = QVBoxLayout(self)
+
+		name_label = QLabel(self._name)
+		name_stretch = QHBoxLayout()
+		name_stretch.addStretch(1)
+		name_stretch.addWidget(name_label)
+		main_layout.addLayout(name_stretch)
+
+		self._exp_layout = QVBoxLayout()
+		self._exp_widget = QFrame()
+		self._exp_widget.setLayout(self._exp_layout)
+
+		main_layout.addWidget(self._exp_widget)
+
+		self.exp_widgets()
+
+		divider = QFrame()
+		divider.setFrameShape(QFrame.HLine)
+		main_layout.addWidget(divider)
+
+		remove = QPushButton("Remove", self)
+		remove.clicked.connect(self.remove)
+
+		hide = QPushButton("Hide", self)
+		hide.clicked.connect(self.hide)
+
+		show = QPushButton("Show", self)
+		show.clicked.connect(self.show)
+
+		stretch = QHBoxLayout()
+		stretch.addStretch(1)
+		stretch.addWidget(hide)
+		stretch.addWidget(show)
+		stretch.addWidget(remove)
+		main_layout.addLayout(stretch)
+
+	def exp_widgets(self):
+
+		pass
+
+	def remove(self):
+
+		 pass
+
+	def hide(self):
+
+		self._exp_widget.hide()
+
+	def show(self):
+
+		self._exp_widget.show()
+
+class LocalExp(Experiments):
+
+	def __init__(self, fitter, exp, name, slider_list, global_var, global_exp, local_exp):
+
+		self._local_exp = local_exp
+		self._exp = exp
+
+		super().__init__(fitter, name, slider_list, global_var, global_exp)
+
+	def exp_widgets(self):
+
+		parameters = self._exp.param_values
+
+		for p, v in parameters.items():
+			s = LocalSliders(self._exp, p, v, self._fitter, self._global_var, self._slider_list, self._global_exp)
+			self._slider_list["Local"][self._exp].append(s)
+			self._exp_layout.addWidget(s)
+
+	def remove(self):
+
+		self._local_exp.pop(self._exp, None)
+		self._slider_list["Local"].pop(self._exp, None)
+		self._fitter.remove_experiment(self._exp)
+		self.close()
+
+class GlobalExp(Experiments):
+
+	def __init__(self, fitter, name, slider_list, global_var, global_exp):
+
+		super().__init__(fitter, name, slider_list, global_var, global_exp)
+
+		self._linked_list = []
+
+	def exp_widgets(self):
+
+		s = GlobalSliders(None, self._name, self._fitter, self._global_exp)
+		self._slider_list["Global"][self._name] = s
+		self._exp_layout.addWidget(s)
+
+	def linked(self, loc_slider):
+
+		self._linked_list.append(loc_slider)
+
+	def unlinked(self, loc_slider):
+
+		self._linked_list.remove(loc_slider)
+
+	def remove(self):
+
+		self._global_exp.pop(self._name, None)
+		self._fitter.remove_global(self._name)
+		self._slider_list["Global"].pop(self._name, None)
+
+		for s in self._linked_list:
+			s.reset()
+
+		self.close()
