--- qtgui/exp_widgets/__init__.py
@@ -1,7 +1,0 @@
-__author__ = "Hiranmayi Duvvuri"
-__date__ = "2017-01-10"
-__all__ = []
-
-
-from .indiv_exp import *
-from .exp_sliders import *
